Pickups and Code Cleanup

+++ Assets/Scripts/Multi-Use/AttackZone.cs
@@ -14,8 +14,8 @@ public class AttackZone : MonoBehaviour
 
     // Attack Variables
     [SerializeField]
-    private int attackDamage;
-    public int damageIncrease;
+    private float attackDamage;
+    public float damageIncrease;
     public Vector2 knockback = Vector2.zero;
 
     // This is a start rather than awake as the 'current' variables for stuff like melee damage are done within the Awake methods, so we want these to be done after
